tests: add functional test
This commit adds a functional test to mimic what the user
would see. It checks that the user data gets written to a json
file and verifies the correct report details are returned.
NOTE: it does not check the repeated interval functionality of `datasaver`.

Refactors were made to `index.js` and `datasaver` to be
able to pass in port numbers so conflicts could be avoided when
running the tests while also having the services running in
development.